requirements.txt
async-generator==1.10
attrs==21.3.0
beautifulsoup4==4.10.0
blinker==1.7.0
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.6
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.5.2
coursera-dl==0.11.5
cryptography==36.0.1
distro==1.7.0
filelock==3.13.3
flask==3.0.3
fsspec==2024.3.1
h11==0.12.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.2.0
Jinja2==3.1.3
Keras-Preprocessing==1.1.2
keyring==23.2.1
llvmlite==0.41.1
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.1
numba==0.58.1
numpy==1.24.4
openai-whisper==20231117
opencv-contrib-python==4.5.3.56
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
outcome==1.1.0
packaging==21.3
pyasn1==0.4.8
pycparser==2.21
pyOpenSSL==21.0.0
pyparsing==3.0.7
pywin32-ctypes==0.2.0
regex==2023.12.25
requests==2.26.0
scikit-build==0.13.1
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sympy==1.12
tiktoken==0.6.0
torch==2.2.2
tqdm==4.66.2
trio==0.19.0
trio-websocket==0.9.2
typing-extensions==4.10.0
urllib3==1.26.7
werkzeug==3.0.2
wsproto==1.0.0
zipp==3.6.0
